requirements.txt
# File watching
watchdog>=4.0.0

# OCR
pytesseract>=0.3.13
Pillow>=10.4.0

# PDF extraction
pdfplumber>=0.11.0
PyMuPDF>=1.24.0

# AI providers
anthropic>=0.30.0
google-generativeai>=0.8.0

# Config
PyYAML>=6.0.2

# Scheduling (batch mode)
schedule>=1.2.2

# Dev / test
pytest>=8.3.0
pytest-cov>=5.0.0
pytest-mock>=3.14.0
